--- css/customize-controls.css
@@ -1,44 +1,0 @@
-/**
- * Styles used for various parts of the customizer.
- *
- * This file incorporates styles from Justin Tadlock's TRT-Customizer-Pro:
- * https://github.com/justintadlock/trt-customizer-pro.
- * Used under GPLv2 licence - 2016 © Justin Tadlock.
- *
- * @package   PattonWebz_Customize
- * @since     1.0.0
- * @author    William Patton <will@pattonwebz.com>
- * @copyright Copyright (c) 2018, William Patton
- * @link      https://github.com/pattonwebz/customizer-framework/
- * @license   http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
- */
-
-#customize-controls .control-section-pattonwebz-customizer-section-help .accordion-section-title:hover,
-#customize-controls .control-section-pattonwebz-customizer-section-help .accordion-section-title:focus {
-	background-color: #fff;
-}
-
-#customize-controls .control-section-pattonwebz-customizer-section-help.active .accordion-section-title {
-	color: #0073aa;
-	background: #f3f3f5;
-	border-left-color: #0073aa;
-}
-
-#customize-theme-controls .control-section-pattonwebz-customizer-section-help .accordion-section-title:after {
-	content: "";
-}
-
-.control-section-pattonwebz-customizer-section-help .accordion-section-title .button {
-	margin-top:  -4px;
-	font-weight: 400;
-	margin-left: 8px;
-}
-
-.rtl .control-section-pattonwebz-customizer-section-help .accordion-section-title .button {
-	margin-left:  0;
-	margin-right: 8px;
-}
-.control-section-pattonwebz-customizer-section-help div.info {
-	background: #f5f5f5;
-	padding: 5px 20px;
-}
