+++ inc/helpers/helpers.php
@@ -3,14 +3,14 @@
  * PattonWebz Customizer Class for helper functions.
  *
  * @package   PattonWebz_Customizer
- * @since     v1.2.0
+ * @since     v2.0.0
  * @author    William Patton <will@pattonwebz.com>
  * @copyright Copyright (c) 2018, William Patton
  * @link      https://github.com/pattonwebz/customizer-framework/
  * @license   http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
  */
 
-namespace Pattonwebz\Customizer;
+namespace PattonWebz\Customizer\Helpers;
 
 /**
  * Class of helper functions.
@@ -24,7 +24,7 @@ class Helpers {
 	public static function get_categories() {
 		// get all the categories.
 		$categories       = get_categories();
-		$categories_array = array();
+		$categories_array = [];
 
 		// Add a new entry to indicate all posts.
 		$categories_array[0] = esc_html__( 'All Categories', 'pattonwebz' );
@@ -33,6 +33,6 @@ class Helpers {
 			$categories_array[ $category->term_id ] = $category->name;
 		}
 		// return an array of available categories.
-		return $categories_array;
+		return apply_filters( 'pattonwebz_customize_filter_categories_list', $categories_array );
 	}
 }
